+++ src/components/App/index.tsx
@@ -1,0 +1,1 @@
+export { App } from './App';
