+++ src/components/article-params-form/ArticleParamsForm.module.scss
@@ -21,6 +21,7 @@
 	height: auto;
 	min-height: 100%;
 	padding: 50px 31px;
+	gap: 50px;
 	background-color: #ffffff;
 	border-right: 1px solid #000000;
 }
