+++ src/ui/separator/index.module.scss
@@ -1,5 +1,5 @@
 .separator {
 	width: 100%;
 	height: 1px;
-	background: #000000;
+	background: #D7D7D7;
 }
